+++ src/commonMain/kotlin/payloads/CreatedComment.kt
@@ -2,7 +2,7 @@ package payloads
 
 @kotlinx.serialization.Serializable
 data class CreatedComment(
-    val timestamp: Double,
+    val timestamp: Int,
     val content: String,
     val attachPrediction: Boolean,
 )
